# api.yml
# This is the OpenAPI specification for your service. It is used to generate the client code.
# To generate the server code, use the Gradle task `openApiGenerate`. The 
# generated code will be placed in `build/generate-resources/main`.
#
# We will evaluate you based on how well you design your API.
# 1. How well it follows REST principles.
# 2. How easy it is to understand and use.
#
# References
# 1. https://swagger.io/specification/
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
  /registration:
    post:
      summary: Registration a new user
      operationId: Registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - fullName
                - password
              properties:
                phoneNumber:
                  type: string
                fullName:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login:
    post:
      summary: Login to the system
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - password
              properties:
                phoneNumber:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /my-profile:
    get:
      summary: Get my profile
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyProfileResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update my profile
      operationId: updateMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phoneNumber:
                  type: string
                fullName:
                  type: string
      responses:
        '200':
          description: My profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateMyProfileResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    RegistrationResponse:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
    LoginResponse:
      type: object
      required:
        - userId
        - jwt
      properties:
        userId:
          type: string
        jwt:
          type: string
    MyProfileResponse:
      type: object
      required:
        - phoneNumber
        - fullName
      properties:
        phoneNumber:
          type: string
        fullName:
          type: string
    UpdateMyProfileResponse:
      type: object
      properties:
        phoneNumber:
          type: string
        fullName:
          type: string
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string